Migrate victory-canvas to TypeScript
