+++ packages/victory-canvas/src/canvas-bar.tsx
@@ -1,7 +1,9 @@
 import { assign } from "lodash";
-import PropTypes from "prop-types";
 import React from "react";
 import {
+  BarProps,
+  VictoryBarAlignmentType,
+  VictoryBarCornerRadiusObject,
   getBarPath,
   getBarWidth,
   getCornerRadius,
@@ -9,27 +11,38 @@ import {
   getStyle,
 } from "victory-bar";
 import { useCanvasContext } from "./hooks/use-canvas-context";
-import { CommonProps } from "victory-core";
+import { NumberOrCallback, VictoryCommonPrimitiveProps } from "victory-core";
 
-const evaluateProps = (props) => {
+export interface CanvasBarProps extends VictoryCommonPrimitiveProps {
+  alignment?: VictoryBarAlignmentType;
+  barOffset?: number[];
+  barRatio?: number;
+  barWidth?: NumberOrCallback;
+  cornerRadius?: NumberOrCallback | VictoryBarCornerRadiusObject;
+  datum?: any;
+  getPath?: (x: number, y: number, size: number) => string;
+  horizontal?: boolean;
+  width?: number;
+  x?: number;
+  y?: number;
+  y0?: number;
+}
+
+const evaluateProps = (props: CanvasBarProps) => {
   /**
    * Potential evaluated props of following must be evaluated in this order:
    * 1) `style`
    * 2) `barWidth`
    * 3) `cornerRadius`
    */
-  const style = getStyle(props.style, props);
+  const style = getStyle(props.style, props as BarProps);
   const barWidth = getBarWidth(props.barWidth, assign({}, props, { style }));
   const cornerRadius = getCornerRadius(
     props.cornerRadius,
     assign({}, props, { style, barWidth }),
   );
-
-  return assign({}, props, {
-    style,
-    barWidth,
-    cornerRadius,
-  });
+  const modifiedProps = assign({}, props, { style, barWidth, cornerRadius });
+  return modifiedProps;
 };
 
 export const usePreviousValue = (value) => {
@@ -40,30 +53,30 @@ export const usePreviousValue = (value) => {
   return ref.current;
 };
 
-const CanvasBar = (initialProps) => {
+export const CanvasBar = (props: CanvasBarProps) => {
   const { canvasRef } = useCanvasContext();
-  const props = evaluateProps(initialProps);
-  const { polar, style, barWidth, cornerRadius, origin } = props;
+  const modifiedProps = evaluateProps(props);
+  const { polar, style, barWidth, cornerRadius, origin } = modifiedProps;
 
   const path2d = React.useMemo(() => {
     const p = polar
-      ? getPolarBarPath(props, cornerRadius)
-      : getBarPath(props, barWidth, cornerRadius);
+      ? getPolarBarPath(modifiedProps, cornerRadius)
+      : getBarPath(modifiedProps, barWidth, cornerRadius);
 
     return new Path2D(p);
-  }, [polar, barWidth, cornerRadius, props]);
+  }, [polar, barWidth, cornerRadius, modifiedProps]);
 
   const previousPath = usePreviousValue(path2d);
 
   const draw = React.useCallback(
-    (ctx, path) => {
+    (ctx: CanvasRenderingContext2D, path: Path2D) => {
       ctx.fillStyle = style.fill;
       ctx.strokeStyle = style.stroke;
       ctx.globalAlpha = style.fillOpacity;
       ctx.lineWidth = style.strokeWidth;
 
       if (polar) {
-        ctx.translate(origin.x, origin.y);
+        ctx.translate(origin?.x || 0, origin?.y || 0);
       }
       ctx.fill(path);
       ctx.setTransform(1, 0, 0, 1, 0, 0);
@@ -73,11 +86,12 @@ const CanvasBar = (initialProps) => {
 
   // This will clear the previous bar without clearing the entire canvas
   const clearPreviousPath = React.useCallback(
-    (ctx) => {
+    (ctx: CanvasRenderingContext2D) => {
       if (previousPath) {
         ctx.save();
         // This ensures that the entire shape is erased
-        ctx.lineWidth = style.strokeWidth + 2;
+        const strokeWidth = (style.strokeWidth as number) || 0;
+        ctx.lineWidth = strokeWidth + 2;
 
         ctx.globalCompositeOperation = "destination-out";
         draw(ctx, previousPath);
@@ -90,8 +104,8 @@ const CanvasBar = (initialProps) => {
   );
 
   React.useEffect(() => {
-    const ctx = canvasRef.current.getContext("2d");
-
+    const ctx = canvasRef.current?.getContext("2d");
+    if (!ctx) return;
     clearPreviousPath(ctx);
     draw(ctx, path2d);
   }, [
@@ -100,38 +114,10 @@ const CanvasBar = (initialProps) => {
     polar,
     barWidth,
     cornerRadius,
-    props,
+    modifiedProps,
     path2d,
     clearPreviousPath,
   ]);
 
   return null;
 };
-
-CanvasBar.propTypes = {
-  ...CommonProps.primitiveProps,
-  alignment: PropTypes.oneOf(["start", "middle", "end"]),
-  barRatio: PropTypes.number,
-  barWidth: PropTypes.oneOfType([PropTypes.number, PropTypes.func]),
-  cornerRadius: PropTypes.oneOfType([
-    PropTypes.number,
-    PropTypes.func,
-    PropTypes.shape({
-      top: PropTypes.oneOfType([PropTypes.number, PropTypes.func]),
-      topLeft: PropTypes.oneOfType([PropTypes.number, PropTypes.func]),
-      topRight: PropTypes.oneOfType([PropTypes.number, PropTypes.func]),
-      bottom: PropTypes.oneOfType([PropTypes.number, PropTypes.func]),
-      bottomLeft: PropTypes.oneOfType([PropTypes.number, PropTypes.func]),
-      bottomRight: PropTypes.oneOfType([PropTypes.number, PropTypes.func]),
-    }),
-  ]),
-  datum: PropTypes.object,
-  getPath: PropTypes.func,
-  horizontal: PropTypes.bool,
-  width: PropTypes.number,
-  x: PropTypes.number,
-  y: PropTypes.number,
-  y0: PropTypes.number,
-};
-
-export default CanvasBar;
