--- packages/victory-canvas/src/index.js
@@ -1,5 +1,0 @@
-export { default as CanvasBar } from "./canvas-bar";
-export { default as CanvasGroup } from "./canvas-group";
-export { default as CanvasCurve } from "./canvas-curve";
-export { default as CanvasPoint } from "./canvas-point";
-export { useCanvasContext } from "./hooks/use-canvas-context";
