+++ packages/victory-canvas/src/index.ts
@@ -1,0 +1,5 @@
+export * from "./canvas-bar";
+export * from "./canvas-group";
+export * from "./canvas-curve";
+export * from "./canvas-point";
+export { useCanvasContext } from "./hooks/use-canvas-context";
